Move S3ObjectLocation into the new storage lib
