+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/message/MessagePointer.scala
@@ -2,7 +2,7 @@ package uk.ac.wellcome.messaging.message
 
 import java.net.URI
 
-import uk.ac.wellcome.s3.S3ObjectLocation
+import uk.ac.wellcome.storage.s3.S3ObjectLocation
 
 import scala.util.Try
