+++ sbt_common/messaging/src/test/scala/uk/ac/wellcome/messaging/message/MessageWriterTest.scala
@@ -6,8 +6,11 @@ import uk.ac.wellcome.messaging.sns.SNSConfig
 import uk.ac.wellcome.messaging.test.fixtures.Messaging
 import uk.ac.wellcome.messaging.test.fixtures.SNS.Topic
 import uk.ac.wellcome.models.aws.S3Config
-import uk.ac.wellcome.storage.s3.{KeyPrefixGenerator, S3ObjectLocation}
-import uk.ac.wellcome.test.fixtures.S3.Bucket
+import uk.ac.wellcome.storage.s3.{
+  KeyPrefixGenerator,
+  S3ObjectLocation
+}
+import uk.ac.wellcome.storage.test.fixtures.S3.Bucket
 import uk.ac.wellcome.utils.JsonUtil._
 
 import scala.util.Success
