+++ sbt_common/storage/src/main/scala/uk/ac/wellcome/storage/vhs/VersionedHybridStore.scala
@@ -6,14 +6,13 @@ import io.circe.{Decoder, Encoder}
 import uk.ac.wellcome.models._
 import uk.ac.wellcome.models.aws.S3Config
 import uk.ac.wellcome.storage.dynamo.{UpdateExpressionGenerator, VersionedDao}
-import uk.ac.wellcome.storage.s3.S3ObjectStore
+import uk.ac.wellcome.storage.s3.{S3ObjectLocation, S3ObjectStore}
 import uk.ac.wellcome.storage.type_classes.{
   HybridRecordEnricher,
   IdGetter,
   VersionGetter,
   VersionUpdater
 }
-import uk.ac.wellcome.s3.S3ObjectLocation
 import uk.ac.wellcome.utils.GlobalExecutionContext._
 
 import scala.concurrent.Future
